+++ src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { RoutesObj } from '../frame/tools/interfaces';
 import { notFound } from './common/not-found.component';
-import { cartPageComponent } from './pages/components/cart-page.component';
-import { homePageComponent } from './pages/components/home-page.component';
+import { cartPageComponent } from './pages/cart-page.component';
+import { homePageComponent } from './pages/home-page.component';
 
 export const appRoutes: RoutesObj[] = [
   { path: '', component: homePageComponent },
