+++ src/app/pages/home-page.component.ts
@@ -1,7 +1,7 @@
-import { DMComponent } from '../../../frame/index';
-import { ComponentConfig } from '../../../frame/tools/interfaces';
-import { filterComponent } from './filter.component';
-import { productListComponent } from './product-list.component';
+import { DMComponent } from '../../frame/index';
+import { ComponentConfig } from '../../frame/tools/interfaces';
+import { filterComponent } from '../components/filter.component';
+import { productListComponent } from '../components/product-list.component';
 
 class HomePageComponent extends DMComponent {
   constructor(config: ComponentConfig) {
